+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 
 ## v1.1.x
 
+* Added: LLT/JBD BLE BMS - Added MAC address as unique identifier. Fixed https://github.com/Louisvdw/dbus-serialbattery/issues/970 by @mr-manuel
 * Added: Venus OS version to logfile by @mr-manuel
 * Changed: Install script now shows repositories and version numbers by @mr-manuel
 * Changed: JKBMS BLE - Fix driver gets unresponsive, if connection is lost https://github.com/Louisvdw/dbus-serialbattery/issues/720 with https://github.com/Louisvdw/dbus-serialbattery/pull/941 by @cupertinomiranda